Support for platform specific dependencies
Substantial changes to accomodate for platform specific dependencies.
Still very rough.
It is incomplete: a '_cfg' attr set needs to added with data from a
'rustc --print cfg' run.
Also, it depends on a little Rust prog(cargo-platform-to-nix) that converts
a 'cfg()' expression to a Nix expression that compares values from a
cfg() expression with the _cfg expression.
Next is a proper conversion to a Nix derivation.